pyproject.toml
[tool.poetry]
name = "Apcer"
version = "0.0.1"
description = ""
authors = ["Kamyab Vijapura <Kamyab.Vijapura@apcerls.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = { extras = ["postgresql-asyncpg"], version = "^2.0.30" }
alembic = "^1.13.1"
fastapi-pagination = "^0.12.13"
fastapi-redis-cache = "^0.2.5"
pydantic-settings = "^2.3.3"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
httpx = "^0.27.0"
apscheduler = "^3.10.1"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic-extra-types = "^2.8.2"
fastapi-limiter = "^0.1.6"
colorama = "^0.4.6"
cryptography = "^45.0.5"
bcrypt = "^4.3.0"
litellm = "^1.74.4"
pymupdf = "^1.26.3"
json-repair = "^0.47.8"
pytz = "^2025.2"
python-ulid = "^3.0.0"
authlib = "^1.6.1"
itsdangerous = "^2.2.0"
celery = "^5.5.3"
sentry-sdk = "^2.34.1"
psutil = "^7.0.0"
phonenumbers = "^9.0.11"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
skip-magic-trailing-comma = true
verbose = true

[tool.ruff]
line-length = 89

[tool.interrogate]
ignore-module = true
ignore-init-method = true
ignore-init-module = true
fail-under = 100
verbose = 2
